# openapi.yaml
openapi: 3.0.3
info:
  title: Telegram Post Generator API
  version: 1.0.0
  description: >
    REST API для управления пользователями, их настройками, постами и автоматической публикацией в Telegram-каналы.
servers:
  - url: http://localhost:4000/api
    description: Локальный сервер разработки
...
paths:
  /users/register:
    post:
      tags: [Users]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, telegram_token, channel_id]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: secret123
                telegram_token:
                  type: string
                  example: 1234567890:ABC-DEF1234ghIklmNOP
                channel_id:
                  type: string
                  example: "@my_channel"
      responses:
        "200":
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  userId: { type: integer }
                  auth_token: { type: string }
        "400":
          description: Ошибка ввода или пользователь уже существует
  /users/login:
    post:
      tags: [Users]
      summary: Вход пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: secret123
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  userId: { type: integer }
                  telegram_token: { type: string }
                  channel_id: { type: string }
                  auth_token: { type: string }
        "403":
          description: Неверный пароль
        "404":
          description: Пользователь не найден
  /users/settings/{userId}:
    get:
      tags: [Users]
      summary: Получение настроек пользователя
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Настройки пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  settings:
                    $ref: "#/components/schemas/UserSettings"
        "403":
          description: Доступ запрещён
        "404":
          description: Пользователь не найден

    put:
      tags: [Users]
      summary: Обновление настроек пользователя
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettings"
      responses:
        "200":
          description: Настройки успешно обновлены
        "403":
          description: Доступ запрещён
        "404":
          description: Пользователь не найден
...
components:
  securitySchemes:
    AuthTokenHeader:
      type: apiKey
      in: header
      name: x-auth-token
      description: >
        Уникальный токен авторизации пользователя.
        Передаётся в каждом запросе, где требуется проверка владельца (в header x-auth-token).
  schemas:
    UserSettings:
      type: object
      properties:
        username:
          type: string
          example: "admin"
        telegram_token:
          type: string
          example: "1234567890:ABC-DEF1234ghIklmNOP"
        channel_id:
          type: string
          example: "@my_channel"
        add_images:
          type: boolean
          example: true
        use_own_posts:
          type: boolean
          example: false
        use_other_channels:
          type: boolean
          example: true
        channels_list:
          type: array
          items: { type: string }
          example: ["@technews", "@funnyvideos"]
        use_sites:
          type: boolean
          example: true
        sites_list:
          type: array
          items: { type: string }
          example: ["https://example.com", "https://another.com"]
...